refactor(bindings/python): move async code to an `asyncio` module
